Refactor codes, add comments
