+++ test/com/linkedin/drelephant/analysis/AnalyticJobTest.java
@@ -18,7 +18,7 @@ package com.linkedin.drelephant.analysis;
 
 import com.linkedin.drelephant.ElephantContext;
 import com.linkedin.drelephant.configurations.heuristic.HeuristicConfigurationData;
-import com.linkedin.drelephant.mapreduce.MapReduceFetcherHadoop2;
+import com.linkedin.drelephant.mapreduce.fetchers.MapReduceFetcherHadoop2;
 import com.linkedin.drelephant.mapreduce.data.MapReduceApplicationData;
 import com.linkedin.drelephant.mapreduce.data.MapReduceCounterData;
 import com.linkedin.drelephant.mapreduce.data.MapReduceTaskData;
